backend/requirements.txt
fastapi==0.110.0
python-multipart==0.0.9
pyjwt==2.8.0
pyyaml==6.0.1
uvicorn==0.27.1
python-jose==3.3.0